# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/getColumns": {
            "post": {
                "description": "accepts and returns json object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "porecess excel file from directus",
                "parameters": [
                    {
                        "description": "req",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DirectusModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FileColumns"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/api/v1/hook": {
            "post": {
                "description": "accepts and returns json object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "porecess excel file from directus",
                "parameters": [
                    {
                        "description": "req",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DirectusModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMsg"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/api/v1/upload/excel": {
            "post": {
                "description": "accept multipart/form-data returns json struct",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Parsing excel file from supplier",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMsg"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.DirectusModel": {
            "type": "object",
            "properties": {
                "accountability": {
                    "type": "object",
                    "properties": {
                        "company": {
                            "type": "string"
                        },
                        "role": {
                            "type": "string"
                        },
                        "user": {
                            "type": "string"
                        }
                    }
                },
                "collection": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                }
            }
        },
        "models.FileColumns": {
            "type": "object",
            "properties": {
                "row_id": {
                    "type": "integer"
                },
                "row_name": {
                    "type": "string"
                }
            }
        },
        "models.ResponseMsg": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}